+++ src/mpservice/mpserver.py
@@ -1299,13 +1299,14 @@ class Server:
             The result is returned to the requester.
             The ``backlog`` value is simply the size limit on this internal
             book-keeping dict.
-
-        .. versionchanged:: 0.13.0
-            The Server object is used in either sync mode or async mode, but not both.
-            Start it in the sync context manager and use it in sync way;
-            start it in the async context manager and use it in async way.
-            In both modes, the main methods are :meth:`call` and :meth:`stream`.
         """
+        if backlog is not None:
+            warnings.warn(
+                "The parameter `backlog` is deprecated in version 0.12.7 and will be removed in 0.14.0. Use `capacity` instead",
+                DeprecationWarning,
+                stacklevel=2,
+            )
+            capacity = backlog
         _init_server(self, servlet=servlet, capacity=capacity, main_cpu=main_cpu)
 
     @property
@@ -1347,10 +1348,10 @@ class Server:
 
     def call(self, x, /, *, timeout: int | float = 60, backpressure: bool = True):
         """
-        The parameters ``x`` and ``timeout`` have the same meanings as in :meth:`AsyncServer.call`.
-
         This method is thread-safe, meaning it can be called from multiple threads
         concurrently.
+
+        .. seealso:: :meth:`AsyncServer.call`
         """
         fut = self._enqueue(x, timeout, backpressure)
         return self._wait_for_result(fut)
@@ -1446,6 +1447,8 @@ class Server:
         This method is thread-safe, that is, multiple threads can call this concurrently
         with their own input streams. In that case, you may want to use ``return_exceptions=True``
         in each of them so that one's exception does not propagate and halt the others.
+
+        It is also fine to have calls to :meth:`call` and :meth:`stream` concurrently.
 
         Parameters
         ----------
@@ -1559,9 +1562,17 @@ class AsyncServer:
         self,
         servlet: Servlet,
         *,
+        backlog: int = None,
         main_cpu: int = 0,
         capacity: int = 256,
     ):
+        if backlog is not None:
+            warnings.warn(
+                "The parameter `backlog` is deprecated in version 0.12.7 and will be removed in 0.14.0. Use `capacity` instead",
+                DeprecationWarning,
+                stacklevel=2,
+            )
+            capacity = backlog
         _init_server(self, servlet=servlet, capacity=capacity, main_cpu=main_cpu)
 
     @property
@@ -1605,8 +1616,6 @@ class AsyncServer:
         """
         When this is called, this server is usually backing a (http or other) service.
         Concurrent async calls to this object may happen.
-        In such use cases, :meth:`call` and :meth:`stream` should not be called to this object
-        at the same time.
 
         Parameters
         ----------
@@ -1730,6 +1739,11 @@ class AsyncServer:
         return_exceptions: bool = False,
         timeout: int | float = 60,
     ) -> AsyncIterator:
+        '''
+        Calls to :meth:`stream` and :meth:`call` can happen at the same time
+        (i.e. interleaved); multiple calls to :meth:`stream` can also happen
+        at the same time by different "users" (in the same thread).
+        '''
         async def _enqueue(tasks, timeout):
             # Putting input data in the queue does not need concurrency.
             # The speed of sequential push is as fast as it can go.