+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [0.13.0] - in progress
 
 - Breaking changes to ``mpserver.Server`` API: the class is split into two classes: ``Server`` and ``AsyncServer``.
+- ``mpserver.Server.call`` got new parameter ``backpressure`` (previously only the async call has this parameter).
 - Finetuned waiting and sleeping logic in ``mpserver.{Server, AsyncServer}``; use ``Condition`` to replace sleeping.
 - Made sure (or confirmed) that ``mpserver.Server._call`` and ``mpserver.Server._stream` are thread-safe.
 - ``streamer.Stream.peek`` finetune of printing; got new parameter ``prefix`` and ``separator``.